+++ TestApp/Sources/Reader/Common/Outline/OutlineTableView.swift
@@ -19,27 +19,28 @@ enum OutlineSection: Int {
 }
 
 struct OutlineTableView: View {
+    private let publication: Publication
     @ObservedObject private var bookmarksModel: BookmarksViewModel
     @ObservedObject private var highlightsModel: HighlightsViewModel
     @State private var selectedSection: OutlineSection = .tableOfContents
     
     // Outlines (list of links) to display for each section.
     private var outlines: [OutlineSection: [(level: Int, link: R2Shared.Link)]] = [:]
-    
+
     init(publication: Publication, bookId: Book.Id, bookmarkRepository: BookmarkRepository, highlightRepository: HighlightRepository) {
-     
+        self.publication = publication
+        self.bookmarksModel = BookmarksViewModel(bookId: bookId, repository: bookmarkRepository)
+        self.highlightsModel = HighlightsViewModel(bookId: bookId, repository: highlightRepository)
+
         func flatten(_ links: [R2Shared.Link], level: Int = 0) -> [(level: Int, link: R2Shared.Link)] {
             return links.flatMap { [(level, $0)] + flatten($0.children, level: level + 1) }
         }
         
-        outlines = [
+        self.outlines = [
             .tableOfContents: flatten(publication.tableOfContents),
             .landmarks: flatten(publication.landmarks),
             .pageList: flatten(publication.pageList)
         ]
-        
-        bookmarksModel = BookmarksViewModel(bookId: bookId, repository: bookmarkRepository)
-        highlightsModel = HighlightsViewModel(bookId: bookId, repository: highlightRepository)
     }
     
     var body: some View {
@@ -55,7 +56,9 @@ struct OutlineTableView: View {
                             .frame(maxWidth: .infinity, alignment: .leading)
                             .contentShape(Rectangle())
                             .onTapGesture {
-                                locatorSubject.send(Locator(link: item.link))
+                                if let locator = publication.locate(item.link) {
+                                    locatorSubject.send(locator)
+                                }
                             }
                     }
                 } else {