+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file. Take a look
 
 ## [Unreleased]
 
+### Deprecated
+
+#### Shared
+
+* `Locator(link: Link)` is deprecated as it may create an incorrect `Locator` if the link `type` is missing.
+    * Use `publication.locate(Link)` instead.
+
 ### Fixed
 
 #### Navigator